add obfuscator package
