+++ package.json
@@ -11,7 +11,7 @@
     "build-orchidui": "npm run build --prefix packages/@orchidui-vue",
     "build-docs": "npm run build --prefix docs",
     "build-playground": "vite build",
-    "build": "npm run build-orchidui && npm run build-docs && npm run build-playground && node encode_class_name.js",
+    "build": "npm run build-orchidui && npm run build-docs && npm run build-playground && node obfuscator.js",
     "build-storybook": "storybook build -o dist/storybook",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix packages",
     "format": "prettier .  --write",
