+++ packages/@orchidui-form/README.md
@@ -1,4 +1,1 @@
-Orchid UI Icons
-
-<!-- command npx @orchidui-icons add-[icon-name] -->
-<!-- command npx @orchidui-icons add-all -->
+# TODO : move form to this package
